requirements.txt
Flask==3.0.3
mediapipe==0.10.14
opencv-python==4.10.0.84
numpy==2.1.1
Pillow==10.4.0
gunicorn==23.0.0
